+++ back/src/forms/resolvers/mutations/markAsProcessed.ts
@@ -1,5 +1,4 @@
-import { EmitterType, Form, Prisma, Status } from "@prisma/client";
-import { prisma } from "@td/prisma";
+import { EmitterType, Prisma, Status } from "@prisma/client";
 
 import { checkIsAuthenticated } from "../../../common/permissions";
 import { MutationResolvers } from "../../../generated/graphql/types";
@@ -21,9 +20,9 @@ import {
   countries,
   isSiret,
   isVat,
-  PROCESSING_OPERATIONS,
-  ProcessingOperationType
+  PROCESSING_OPERATIONS
 } from "@td/constants";
+import { operationHooksQueue } from "../../../queue/producers/bsdUpdate";
 
 const markAsProcessedResolver: MutationResolvers["markAsProcessed"] = async (
   parent,
@@ -127,79 +126,12 @@ const markAsProcessedResolver: MutationResolvers["markAsProcessed"] = async (
     return processedForm;
   });
 
-  await operationHook(processedForm, processedForm.id);
+  await operationHooksQueue.add({
+    operation: processedForm,
+    formId: processedForm.id
+  });
 
   return getAndExpandFormFromDb(processedForm.id);
 };
 
 export default markAsProcessedResolver;
-
-type OperationHookArgs = Pick<
-  Form,
-  | "readableId"
-  | "quantityReceived"
-  | "processingOperationDone"
-  | "recipientCompanySiret"
-  | "recipientCompanyName"
-  | "noTraceability"
->;
-
-// Hook qui est appelé à chaque fois qu'un applique une opération
-// de traitement sur un bordereau
-// À ajouter dans une queue plutôt que de faire ça en synchrone
-export async function operationHook(
-  // Informations sur le traitement final
-  operation: OperationHookArgs,
-  // Identifiant du bordereau
-  formId: string
-) {
-  // Codes de traitement finaux
-  const finalOperationCodes = PROCESSING_OPERATIONS.filter(
-    p =>
-    p.type === ProcessingOperationType.Eliminiation ||
-    p.type === ProcessingOperationType.Valorisation
-    ).map(p => p.code);
-    if (
-    // Le code n'est appelé qu'en cas de traitement final ou de rupture de traçabilité
-    finalOperationCodes.includes(operation.processingOperationDone!) ||
-    operation.noTraceability
-  ) {
-    // On va chercher tous les bordereaux initiaux
-    const formWithInitialForms = await prisma.form.findUniqueOrThrow({
-      where: { id: formId, isDeleted: false },
-      include: {
-        forwarding: true,
-        grouping: { include: { initialForm: true } }
-      }
-    });
-
-    const initialForms = [
-      formWithInitialForms.forwarding,
-      ...(formWithInitialForms.grouping ?? []).map(g => g.initialForm)
-    ].filter(Boolean);
-
-    for (const initialForm of initialForms) {
-      await prisma.form.update({
-        where: { id: initialForm.id },
-        data: {
-          finalOperations: {
-            create: {
-              finalBsdReadableId: operation.readableId,
-              quantity: operation.quantityReceived!,
-              operationCode: operation.processingOperationDone!,
-              destinationCompanySiret: operation.recipientCompanySiret!,
-              destinationCompanyName: operation.recipientCompanyName!
-            }
-          }
-        }
-      });
-
-      // Applique le hook de façon récursive sur les bordereaux initiaux
-      // TODO gérer ça avec une queue
-      // Potentiel point particulier : s'il y a eu scission puis regroupement
-      // le hook pourrait être appelé plusieurs fois, il faudrait donc ajouter
-      // un check sur l'unicité de finalBsdReadableId
-      await operationHook(operation, initialForm.id);
-    }
-  }
-}
