+++ back/src/queue/jobs/index.ts
@@ -1,3 +1,4 @@
 export { sendMailJob } from "./sendmail";
 export { indexBsdJob } from "./indexBsd";
+export { operationHookJob } from "./operationHook";
 export { sendHookJob } from "./sendHook";
