+++ back/src/queue/producers/bsdUpdate.ts
@@ -1,5 +1,5 @@
 import Queue from "bull";
-
+import { OperationHookArgs } from "../jobs/operationHook";
 const { REDIS_URL, NODE_ENV } = process.env;
 
 export type BsdUpdateQueueItem = {
@@ -17,6 +17,17 @@ export const updatesQueue = new Queue<BsdUpdateQueueItem>(
   {
     defaultJobOptions: {
       delay: INDEX_REFRESH_INTERVAL, // We delay processing to make sure updates have been refreshed in ES
+      removeOnComplete: 100
+    }
+  }
+);
+
+// Updates queue, used by the notifier. Items are enqueued once indexation is done
+export const operationHooksQueue = new Queue<OperationHookArgs>(
+  `queue_operation_hook_${NODE_ENV}`,
+  REDIS_URL!,
+  {
+    defaultJobOptions: {
       removeOnComplete: 100
     }
   }
